Close modal with Esc key
What did not solve the problem: xpath - no because there are 2 of them (adding to the list and clicking the first one did not help); css selector - correct, but Selenium did not find it; clicking on `body` by Selenium and by JS; clicking on header.
Moving the cursor worked, but was risky - not sure where the user would leave it.
Additional bug: adding closing modal to `speedtest.py` - because `ec.visibility_of_element_located` is in `screenshot.py`.